+++ frontend/packages/kubevirt-plugin/src/utils/url.ts
@@ -1,4 +1,5 @@
 import { k8sBasePath } from '@console/internal/module/k8s';
+import { VMWizardName, VMWizardMode } from '../constants/vm';
 
 const ELLIPSIS = '…';
 
@@ -45,3 +46,20 @@ export const resolveURL = ({ urlObj, maxHostnameParts, maxPathnameParts }) =>
   urlObj.origin === 'null'
     ? urlObj.href
     : `${resolveOrigin(urlObj, maxHostnameParts)}${resolvePathname(urlObj, maxPathnameParts)}`;
+
+export const getVMWizardCreateLink = (
+  namespace: string,
+  itemName: VMWizardName = VMWizardName.WIZARD,
+  mode: VMWizardMode = VMWizardMode.TEMPLATE,
+  template?: string,
+) => {
+  // Overide mode if name is import.
+  const wizardMode = itemName === VMWizardName.IMPORT ? VMWizardMode.IMPORT : mode;
+  const type = itemName === VMWizardName.YAML ? '~new' : '~new-wizard';
+
+  let params = new URLSearchParams();
+  params.append('mode', wizardMode);
+  params.append('template', template);
+
+  return `/k8s/ns/${namespace || 'default'}/virtualization/${type}${params ? `?${params}` : ''}`; // covers 'yaml', new-wizard and default
+};
